requirements.txt
discord-py-interactions==5.0.0
python-dotenv==0.19.1
